requirements.txt
asttokens==3.0.0
brotli
certifi
cffi
charset-normalizer
colorama==0.4.6
comm==0.2.2
contourpy
cycler
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fonttools
h2
hpack
hyperframe
idna
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver
littleutils
llvmlite==0.44.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath
munkres==1.1.4
nest-asyncio==1.6.0
numba==0.61.0
numpy==2.1.3
outdated
packaging
pandas
pandas_flavor
parso==0.8.4
patsy
pillow
pingouin
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pycparser
Pygments==2.19.1
pynndescent==0.5.13
pyparsing
PySide6==6.8.2
PySocks
python-dateutil
pytz
pywin32==309; sys_platform == "win32"
pyzmq==26.2.1
requests
scikit-learn==1.6.1
scikit-posthocs
scipy==1.15.2
seaborn
setuptools==75.8.2
shiboken6==6.8.2
six
stack-data==0.6.3
statsmodels
tabulate
threadpoolctl
tornado
tqdm==4.67.1
traitlets==5.14.3
tzdata
umap-learn==0.5.7
urllib3
wcwidth==0.2.13
win_inet_pton; sys_platform == "win32"
xarray
zstandard==0.23.0

